Bug fix: Fetch failures 0 members: KeyError in convert_row() Follow-up from #836.
The cause of the bug is that a certain field was missing from the enclave object returned. Normally during convert_row(), if this happens, the KeyError is skipped. However, convert_row() behaves differently now when it is used by DB update functions, and I hadn't accounted for skipping past this KeyError in that case.