- Added ability to turn of api call logging:   - api_logger.py: API_CALL_LOGGING_ON=False   - which makes it so next-api-call-group-id returns null
Because was having problems when our auth token got hosed. Most of
the calls to the backend from the front don't require a connection
to the enclave, except for who-am-i. I was going to set it so if
call logging was turned off, I wouldn't call whoami. But now that
the token and whoami seem to be working ok, don't need to do that. But
I'm going to leave logging turned off just because we're not using it
lately and it complicates the code.